backend/requirements.txt
# Portley Backend Requirements
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Supabase
supabase==2.3.4
supabase-auth==0.1.0

# Rate limiting
slowapi==0.1.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
resend==0.8.0

# AI
google-generativeai==0.3.0

# Payments
razorpay==1.4.1
stripe==7.8.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Utilities
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
